refactor(controllers): dedupe waitForStatus and split HR valuesFrom helpers
Move identical waitForStatus polling helper from three controller test
packages into testutil.WaitForStatus, and extract the HelmRelease
valuesFrom omission cluster (~168 LOC) from controller.go into a new
valuesfrom.go, mirroring kustomization's dispatch/substitute split.
No behavioral change.
